% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_objects.R
\name{AsyncRecognizeRequest}
\alias{AsyncRecognizeRequest}
\title{AsyncRecognizeRequest Object}
\usage{
AsyncRecognizeRequest(audio = NULL, config = NULL)
}
\arguments{
\item{audio}{[Required] The audio data to be recognized}

\item{config}{[Required] The `config` message provides information to the recognizer}
}
\value{
AsyncRecognizeRequest object
}
\description{
AsyncRecognizeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
`AsyncRecognizeRequest` is the top-level message sent by the client forthe `AsyncRecognize` method.
}
\seealso{
Other AsyncRecognizeRequest functions: \code{\link{asyncrecognize}}
}

