% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_objects.R
\name{RecognitionAudio}
\alias{RecognitionAudio}
\title{RecognitionAudio Object}
\usage{
RecognitionAudio(uri = NULL, content = NULL)
}
\arguments{
\item{uri}{URI that points to a file that contains audio data bytes as specified in}

\item{content}{The audio data bytes encoded as specified in}
}
\value{
RecognitionAudio object
}
\description{
RecognitionAudio Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains audio data in the encoding specified in the `RecognitionConfig`.Either `content` or `uri` must be supplied. Supplying both or neitherreturns google.rpc.Code.INVALID_ARGUMENT. See[audio limits](https://cloud.google.com/speech/limits#content).
}
