% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_objects.R
\name{RecognitionConfig}
\alias{RecognitionConfig}
\title{RecognitionConfig Object}
\usage{
RecognitionConfig(maxAlternatives = NULL, languageCode = NULL,
  speechContext = NULL, encoding = NULL, profanityFilter = NULL,
  sampleRate = NULL, audioChannels = NULL)
}
\arguments{
\item{maxAlternatives}{[Optional] Maximum number of recognition hypotheses to be returned}

\item{languageCode}{[Optional] The language of the supplied audio as a BCP-47 language tag}

\item{speechContext}{[Optional] A means to provide context to assist the speech recognition}

\item{encoding}{[Required] Encoding of audio data sent in all `RecognitionAudio` messages}

\item{profanityFilter}{[Optional] If set to `true`, the server will attempt to filter out}

\item{sampleRate}{[Required] Sample rate in Hertz of the audio data sent in all}

\item{audioChannels}{[Optional] The number of interleaved channels in the input audio data}
}
\value{
RecognitionConfig object
}
\description{
RecognitionConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The `RecognitionConfig` message provides information to the recognizerthat specifies how to process the request.
}

