% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_functions.R
\docType{package}
\name{speech_googleAuthR}
\alias{speech_googleAuthR}
\alias{speech_googleAuthR-package}
\title{Google Cloud Speech API
Google Cloud Speech API.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:47:37
filename: /Users/mark/dev/R/autoGoogleAPI/googlespeechv1beta1.auto/R/speech_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}
}

