% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_objects.R
\name{SpeechContext}
\alias{SpeechContext}
\title{SpeechContext Object}
\usage{
SpeechContext(phrases = NULL)
}
\arguments{
\item{phrases}{[Optional] A list of strings containing words and phrases 'hints' so that}
}
\value{
SpeechContext object
}
\description{
SpeechContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Provides 'hints' to the speech recognizer to favor specific words and phrasesin the results.
}

