% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_objects.R
\name{SpeechRecognitionAlternative}
\alias{SpeechRecognitionAlternative}
\title{SpeechRecognitionAlternative Object}
\usage{
SpeechRecognitionAlternative(transcript = NULL, confidence = NULL)
}
\arguments{
\item{transcript}{[Output-only] Transcript text representing the words that the user spoke}

\item{confidence}{[Output-only] The confidence estimate between 0}
}
\value{
SpeechRecognitionAlternative object
}
\description{
SpeechRecognitionAlternative Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Alternative hypotheses (a.k.a. n-best list).
}

