% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_objects.R
\name{SpeechRecognitionResult}
\alias{SpeechRecognitionResult}
\title{SpeechRecognitionResult Object}
\usage{
SpeechRecognitionResult(alternatives = NULL)
}
\arguments{
\item{alternatives}{[Output-only] May contain one or more recognition hypotheses (up to the}
}
\value{
SpeechRecognitionResult object
}
\description{
SpeechRecognitionResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A speech recognition result corresponding to a portion of the audio.
}

