% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_functions.R
\name{syncrecognize}
\alias{syncrecognize}
\title{Perform synchronous speech-recognition: receive results after all audiohas been sent and processed.}
\usage{
syncrecognize(SyncRecognizeRequest)
}
\arguments{
\item{SyncRecognizeRequest}{The \link{SyncRecognizeRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/speech/}{Google Documentation}

Other SyncRecognizeRequest functions: \code{\link{SyncRecognizeRequest}}
}

