% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_objects.R
\name{SyncRecognizeResponse}
\alias{SyncRecognizeResponse}
\title{SyncRecognizeResponse Object}
\usage{
SyncRecognizeResponse(results = NULL)
}
\arguments{
\item{results}{[Output-only] Sequential list of transcription results corresponding to}
}
\value{
SyncRecognizeResponse object
}
\description{
SyncRecognizeResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
`SyncRecognizeResponse` is the only message returned to the client by`SyncRecognize`. It contains the result as zero or more sequential`SpeechRecognitionResult` messages.
}

