% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_objects.R
\name{AsyncRecognizeRequest}
\alias{AsyncRecognizeRequest}
\title{AsyncRecognizeRequest Object}
\usage{
AsyncRecognizeRequest(config = NULL, audio = NULL)
}
\arguments{
\item{config}{*Required* Provides information to the recognizer that specifies how to}

\item{audio}{*Required* The audio data to be recognized}
}
\value{
AsyncRecognizeRequest object
}
\description{
AsyncRecognizeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The top-level message sent by the client for the `AsyncRecognize` method.
}
\seealso{
Other AsyncRecognizeRequest functions: \code{\link{asyncrecognize}}
}
