% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_objects.R
\name{RecognitionConfig}
\alias{RecognitionConfig}
\title{RecognitionConfig Object}
\usage{
RecognitionConfig(maxAlternatives = NULL, sampleRate = NULL,
  languageCode = NULL, speechContext = NULL, encoding = NULL,
  profanityFilter = NULL)
}
\arguments{
\item{maxAlternatives}{*Optional* Maximum number of recognition hypotheses to be returned}

\item{sampleRate}{*Required* Sample rate in Hertz of the audio data sent in all}

\item{languageCode}{*Optional* The language of the supplied audio as a BCP-47 language tag}

\item{speechContext}{*Optional* A means to provide context to assist the speech recognition}

\item{encoding}{*Required* Encoding of audio data sent in all `RecognitionAudio` messages}

\item{profanityFilter}{*Optional* If set to `true`, the server will attempt to filter out}
}
\value{
RecognitionConfig object
}
\description{
RecognitionConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Provides information to the recognizer that specifies how to process therequest.
}
