% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_objects.R
\name{SpeechRecognitionAlternative}
\alias{SpeechRecognitionAlternative}
\title{Google Cloud Speech API Objects 
Google Cloud Speech API.}
\usage{
SpeechRecognitionAlternative(confidence = NULL, transcript = NULL)
}
\arguments{
\item{confidence}{*Output-only* The confidence estimate between 0}

\item{transcript}{*Output-only* Transcript text representing the words that the user spoke}
}
\value{
SpeechRecognitionAlternative object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:15:48
filename: /Users/mark/dev/R/autoGoogleAPI/googlespeechv1beta1.auto/R/speech_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
SpeechRecognitionAlternative Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Alternative hypotheses (a.k.a. n-best list).
}
