% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_objects.R
\name{SyncRecognizeRequest}
\alias{SyncRecognizeRequest}
\title{SyncRecognizeRequest Object}
\usage{
SyncRecognizeRequest(audio = NULL, config = NULL)
}
\arguments{
\item{audio}{[Required] The audio data to be recognized}

\item{config}{[Required] The `config` message provides information to the recognizer}
}
\value{
SyncRecognizeRequest object
}
\description{
SyncRecognizeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
`SyncRecognizeRequest` is the top-level message sent by the client forthe `SyncRecognize` method.
}
\seealso{
Other SyncRecognizeRequest functions: \code{\link{syncrecognize}}
}

